models/knowledge-api/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2

# Neo4j
neo4j==5.15.0

# Vector Search
faiss-cpu==1.7.4
numpy>=1.24.0

# HTTP Client
httpx==0.25.2

# OpenAI (for embeddings)
openai>=1.3.0
